LPS-54970 combine SassExecutorUtil into SassToCssBuilder
