+++ portal-impl/src/com/liferay/portal/tools/sass/SassString.java
@@ -16,6 +16,7 @@ package com.liferay.portal.tools.sass;
 
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.servlet.filters.dynamiccss.RTLCSSUtil;
+import com.liferay.portal.tools.SassToCssBuilder;
 import com.liferay.portal.util.CSSBuilderUtil;
 
 /**
@@ -25,7 +26,7 @@ import com.liferay.portal.util.CSSBuilderUtil;
 public class SassString implements SassFragment {
 
 	public SassString(String fileName, String sassContent) throws Exception {
-		String cssContent = SassExecutorUtil.parse(
+		String cssContent = SassToCssBuilder.parse(
 			fileName, CSSBuilderUtil.parseStaticTokens(sassContent));
 
 		if (fileName.contains("_rtl")) {
